requirements.txt
# Requisitos del proyecto 
pytest 
Pillow 
fastapi
uvicorn
websockets 